+++ ceammc/ext/src/synth/mod_synth.cpp
@@ -1,4 +1,7 @@
 #include "mod_synth.h"
+#include "ceammc_platform.h"
+#include "synth_bthree.h"
+#include "synth_wurley.h"
 
 extern "C" void setup_synth0x2erisset_tone_tilde();
 extern "C" void setup_synth0x2erisset_arp_tilde();
@@ -19,4 +22,7 @@ void ceammc_synth_setup()
     setup_synth0x2eeguitar_tilde();
     setup_synth0x2echurch_bell_tilde();
     setup_synth0x2edubdub_tilde();
+
+    setup_synth_wurley();
+    setup_synth_bthree();
 }