+++ ceammc/ext/src/synth/stksynth.cpp
@@ -1,0 +1,145 @@
+#include "ADSR.h"
+#include "Instrmnt.h"
+#include "SineWave.h"
+
+#include "ceammc_convert.h"
+#include "stksynth.h"
+//#include "synth_bthree.h"
+//#include "synth_wurley.h"
+
+#include "stksynth_p.h"
+
+static t_symbol* SYM_CC = gensym("cc");
+
+StkSynth::StkSynth(const PdArgs& args, stk::Instrmnt* instr)
+    : SoundExternal(args)
+    , synth_(instr)
+    , freq_(nullptr)
+    , gate_(0)
+{
+    freq_ = new FloatPropertyMin("@freq", mtof(48), 0);
+    createProperty(freq_);
+
+    createCbProperty("@gate", &StkSynth::propGate, &StkSynth::propSetGate);
+
+    createSignalOutlet();
+}
+
+StkSynth::~StkSynth()
+{
+    delete synth_;
+}
+
+void StkSynth::setupDSP(t_signal** sp)
+{
+    SoundExternal::setupDSP(sp);
+    stk::Stk::setSampleRate(samplerate());
+}
+
+void StkSynth::processBlock(const t_sample**, t_sample** out)
+{
+    const size_t BS = blockSize();
+
+    for (size_t i = 0; i < BS; i++)
+        out[0][i] = synth_->tick();
+}
+
+AtomList StkSynth::propGate() const
+{
+    return Atom(gate_);
+}
+
+void StkSynth::propSetGate(const AtomList& lst)
+{
+    if (!checkArgs(lst, ARG_FLOAT)) {
+        OBJ_ERR << "gate value expected: " << lst;
+        return;
+    }
+
+    t_float amp = ceammc::clip<t_float>(lst.floatAt(0, 60), 0, 1);
+
+    if (amp > 0)
+        synth_->noteOn(freq_->value(), amp);
+    else
+        synth_->noteOff(freq_->value());
+
+    gate_ = amp;
+}
+
+void StkSynth::m_cc(t_symbol* s, const AtomList& lst)
+{
+    if (!checkArgs(lst, ARG_INT, ARG_FLOAT, s))
+        return;
+
+    int ctl_num = lst[0].asInt();
+    stk::StkFloat ctl_val = lst[1].asFloat();
+
+    synth_->controlChange(ctl_num, ctl_val);
+}
+
+t_float StkSynth::getCC(int n) const
+{
+    // it's a hack
+    //    typedef StkFMSynth<SynthBThree> SB3;
+    //    typedef StkFMSynth<SynthWurley> SWurley;
+
+#define CAST(T) dynamic_cast<const T*>(synth_)
+#define FM_CC(T, n)                              \
+    {                                            \
+        if (CAST(T))                             \
+            return CAST(T)->getControlChange(n); \
+    }
+
+    //    FM_CC(SB3, n);
+    //    FM_CC(SWurley, n);
+
+#undef CAST
+
+    return -1;
+}
+
+class FreqGetter : public stk::SineWave {
+public:
+    FreqGetter(const stk::SineWave* ptr)
+        : stk::SineWave(*ptr)
+    {
+    }
+
+    t_float freq() const
+    {
+        return (rate_ * stk::Stk::sampleRate()) / TABLE_SIZE;
+    }
+};
+
+class ADSRGetter : public stk::ADSR {
+public:
+    ADSRGetter(const stk::ADSR* ptr)
+        : stk::ADSR(*ptr)
+    {
+    }
+
+    t_float target() const
+    {
+        return target_;
+    }
+};
+
+SinWaveFreq::SinWaveFreq(const stk::SineWave* ptr)
+    : ptr_(ptr)
+{
+}
+
+t_float SinWaveFreq::getFrequency() const
+{
+    return FreqGetter(ptr_).freq();
+}
+
+ADSRProp::ADSRProp(const stk::ADSR* ptr)
+    : ptr_(ptr)
+{
+}
+
+t_float ADSRProp::getTarget() const
+{
+    return ADSRGetter(ptr_).target();
+}